requirements.txt
alembic==1.16.4
asyncpg==0.30.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
greenlet==3.2.3
httpx==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==8.4.1
pytest-postgresql==7.0.2
ruff==0.12.4
SQLAlchemy==2.0.41
sqlmodel==0.0.24
sqlparse==0.4.4
starlette==0.47.1
uvicorn==0.35.0
